Abstracted Footer from Dashboard, refactored App
