+++ src/App.js
@@ -1,20 +1,25 @@
 import React from 'react';
-import { Router, Redirect, navigate } from '@reach/router';
-import ChooseYourContact from './components/ChooseYourContact';
+import { Router } from '@reach/router';
+
 import { firebase, db } from './firebase';
 import Login from './components/Auth/Login';
 import DashBoard from './components/DashBoard';
+import LandingPage from './components/LandingPage';
+import SignUp from './components/Auth/SignUp';
+
+import NavBar from './components/NavBar';
+import DashMain from './components/DashMain';
+import Footer from './components/Footer';
+
+import ChooseCallPlan from './components/ChooseCallPlan';
+import ChooseYourContact from './components/ChooseYourContact';
 import ScheduleFreeCall from './components/ScheduleFreeCall';
 import SchedulePaidCall from './components/SchedulePaidCall';
-import ChooseCallPlan from './components/ChooseCallPlan';
 import CallConfirmation from './components/CallConfirmation';
-import AboutUs from './components/AboutUs';
 import PreviousCalls from './components/dashboard/PreviousCalls';
-import NavBar from './components/NavBar';
-import SignUp from './components/Auth/SignUp';
-import MobileLandingPage from './components/MobileLandingPage';
-
+import AboutUs from './components/AboutUs';
 import UpdateAccount from './components/UpdateAccount';
+
 import { fetchUser } from './app/utils';
 
 function useAuth() {
@@ -57,31 +62,30 @@ function useAuth() {
 }
 
 function App() {
-  const result = firebase.auth().onAuthStateChanged(user => {
-    if (user) {
-      navigate(`/user/${user.uid}`);
-    } else {
-      navigate('/');
-    }
-  });
-
-  return (
+  const user = useAuth();
+  console.log(user, 'USERRRRR');
+  return user ? (
     <>
-      {/* <NavBar user={user} /> */}
+      <NavBar user={user} />
       <Router>
-        {/* <Redirect from='/' to={`user/${user.uid}`} noThrow /> */}
-        <DashBoard user={result.user} path='/user/:userId' />
-        <ChooseYourContact user={result.user} path='/choose/:userId' />
+        <DashMain user={user} path='/' exact />
+        <ChooseYourContact user={user} path='/choose/:userId' />
         <ChooseCallPlan path='/choose/:userId/:contactId/call-plan' />
         <ScheduleFreeCall path='/choose/:userId/:contactId/:frequency/schedule-free' />
         <SchedulePaidCall path='/choose/:userId/:contactId/:frequency/schedule' />
         <CallConfirmation path='/confirmation/:contactId' />
         <PreviousCalls path='prev-calls/:userId' />
         <AboutUs path='/about-us' />
-        <UpdateAccount user={result.user} path='/account/:userId' />
-        <Login path='/login' />
+        <UpdateAccount user={user} path='/account/:userId' />
+      </Router>
+      <Footer />
+    </>
+  ) : (
+    <>
+      <Router>
+        <LandingPage path='/' exact />
         <SignUp path='/signup' />
-        <MobileLandingPage path='/' exact />
+        <Login path='/login' />
       </Router>
     </>
   );
